+++ kwhen/src/commonMain/kotlin/ap/panini/kwhen/common/parsers/ParserGenericNumbers.kt
@@ -8,6 +8,6 @@ internal open class ParserGenericNumbers(override val config: Config) : ParserBy
         get() = "\\d+".toRegex()
 
     override fun onMatch(match: MatchResult): DateTime {
-        return DateTime(generalNumber = match.value.toInt())
+        return DateTime(generalNumber = match.value.toDouble())
     }
 }
